Insider Threats in Verizon
The greatest challenge is to figure out who the insider threats are, why they are a threat to businesses and other domains, how can business insider threats remain undetectable for long periods of time, what are the common threats presented by a insider threat, how are insider threats finally detected, what methods were used to detect them, unknowingly evidence left behind in physical and information systems within domains. Detecting and preventing insider threats are the two major challenges in businesses. The teams that work in information systems find it challenging to detect and reveal insider threats, because there is a lack of data and other evidence to prove who the insider threat is.