Business Customer Segmentation
Customers are the most important people to businesses that rely on their purchases to maintain their existence in the market. Without customers, they cannot grow, cannot profit, and they cannot be a business without customers purchasing their products as services. Businesses need money to grow. The problems that they face are unsatisfied consumers, and lack of demand for some products or the majority of products. Customer segmentation is an important topic among businesses and stakeholders in a data science field. It is one of the most vital applications within unsupervised machine learning. 
The clustering method detected multiple segment groups of customers targeted for future improbable user-based. The K-means clustering was used in the project, and it's possessive and essential algorithm 4 / touring dataset that lack labels.